backend/requirements/base.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
pydantic==2.6.1
pydantic-settings==2.1.0
python-multipart==0.0.9
fastapi-cache2==0.2.1
redis==5.0.1  # We will use Redis later
slowapi==0.1.9
sqlalchemy==2.0.46
aiosqlite==0.22.1
